requirements.txt
astroid==2.2.5
awscli==1.16.253
botocore==1.12.243
chardet==3.0.4
colorama==0.4.1
django==2.2.4
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
docutils==0.15.2
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pyasn1==0.4.7
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.24.2
wrapt==1.11.2
django-storages==1.7
boto3==1.9